requirements.txt
Django==1.8.2
Fabric==1.10.1
django-bootstrap3==6.1.0
django-model-utils==2.2
python-coveralls==2.5.0
psycopg2==2.6.1
python-social-auth==0.2.11
# for better django shell_plus
django-extensions==1.5.5
djangorestframework==3.1.3
django-extra-views==0.7.1
raven==5.3.1
ipython
django-debug-toolbar
pytest
hachoir-metadata
hachoir-core
hachoir-parser
